% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{bed2Seq}
\alias{bed2Seq}
\title{Fetch sequences from bed-formatted data frames}
\usage{
bed2Seq(bedline, offset)
}
\arguments{
\item{bedline}{Valid bed-formatted data frame}

\item{offset}{Number of nucleotides expanded on each side ([1,1000])}
}
\value{
A DNAStringset containing the sequences in hg19 genome assembly
}
\description{
Fetch sequences from bed-formatted data frames
}
