\name{cTF}
\alias{cTF}
\docType{data}
\title{Conserved transcription factors (reference data)
}
\description{
Tab-delimited data containing the 634 conserved transcription factors among mouse-rat-human alignments. The list is obtained from the tfbsConsFactors table of UCSC Table Browser.   
}
\format{
  \describe{
    \item{\code{name}}{Identifier of the conserved transcription factor}
    \item{\code{factor}}{Name of the transcription factor}
}
  }
\keyword{datasets}
