% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{filterByCGI}
\alias{filterByCGI}
\title{Filter by CpG islands}
\usage{
filterByCGI(input.data)
}
\arguments{
\item{input.data}{A data frame containing input data in bed format}
}
\value{
A data frame with the CpG sites located in CpG islands
}
\description{
Selects genomic coordinates included in CpG islands, using the cpgIslands dataset.
}
