% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{filterByProm}
\alias{filterByProm}
\title{Filter by promoters}
\usage{
filterByProm(input.data, up.tss, down.tss)
}
\arguments{
\item{input.data}{A data frame containing input data in bed format}

\item{up.tss}{Number of nucleotides upstream transcription start site}

\item{down.tss}{Number of nucleotides downstream transcription start site}
}
\value{
A data frame with the CpG sites located in promoter regions
}
\description{
Selects genomic coordinates than are included in promoters based on the UCSC hg19 gene coordinates.
}
