% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splicing.R
\name{findAltSplicing}
\alias{findAltSplicing}
\title{Find alternative splicing events}
\usage{
findAltSplicing(bed.data, known.alt.splice = NULL)
}
\arguments{
\item{bed.data}{A data frame containing input bed-formatted data}

\item{known.alt.splice}{(optional) Full local path to the UCSC knownalt table.
If the table is not available locally then the script will fetch known alternative splicing
events from UCSC (needs Internet connection).}
}
\value{
1/ A data frame with the identified alternative splicing event overlaps (hg19)

2/ A summary table with the frequency of each alternative splicing event compared to the
reference frequency

3/ A data frame with the number of alternative splicing events per sequence (input to \code{meinter} function)

4/ An overlayed bar chart object
}
\description{
Idetifies known alternative splicing events co-localized with input data.
}
