% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transf.R
\name{findConservedTFBS}
\alias{findConservedTFBS}
\title{Find differentially methylated sites overlapping human/mouse/rat conserved transcription factor binding sites.}
\usage{
findConservedTFBS(bed.data, known.conserved.tfbs.file = NULL)
}
\arguments{
\item{bed.data}{A data frame containing input bed-formatted data}

\item{known.conserved.tfbs.file}{(optional) Full local path to the UCSC conserved transcription factor binding sites.
If the table is not available locally then the script will fetch it from UCSC (needs Internet connection).
NOTE: It is recommended to download the compressed file (Unzipped file >290MB)}
}
\value{
1/ Data frame containing overlaps between \code{bed.data} and conserved transcription factor binding sites

2/ Frequency table of conserved transcription factors on human genome (input to \code{scatterConsTF} function)

3/ A data frame with the number of conserved transcription factor binding sites per sequence (input to \code{meinter} function)
}
\description{
Detects transcription factor binding sites that are conserved in human/mouse/rat alignments and overlap with the input data.
A binding site is considered to be conserved across the alignment if its score meets the threshold score for its binding
matrix in all three species. The score and threshold are computed with the Transfac Matrix
Database (v7.0) created by Biobase. The data are purely computational, and as such not all binding sites listed here are
biologically functional binding sites.
}
