% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quadromes.R
\name{findPals}
\alias{findPals}
\title{Find palindromes in a bed-formatted dataset}
\usage{
findPals(bed.data, offset = 10, min.arm = 5, max.loop = 5,
  max.mismatch = 1)
}
\arguments{
\item{bed.data}{A data frame containing input bed-formatted data}

\item{offset}{Number of nucleotides expanded on each side (default:10, max:200)}

\item{min.arm}{Minimum length of each arm (default:5)}

\item{max.loop}{Maximum length of the loop between the two arms of the palindrome}

\item{max.mismatch}{The maximum number of mismatching letters
allowed between the two arms of the palindromes}
}
\value{
1/ DNAString subject with the identified palindromes

2/ Number of palindromes falling on/neighboring input data

3/ Number of palindromes per sequence (input to `meinter` function)
}
\description{
Deetects whether the target cytosine overlaps with a palindromic sequences or it is located inbetween of the two
arms of a palindromic sequence i.e. in the loop formed by the palindrome.
}
