% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splicing.R
\name{findSpliceSites}
\alias{findSpliceSites}
\title{Find splice sites}
\usage{
findSpliceSites(bed.data, persim = 0.8, offset = 10)
}
\arguments{
\item{bed.data}{A data frame containing input bed-formatted data}

\item{persim}{Similarity with the splice site consensus (default:0.8, range between [0,1])}

\item{offset}{Number of nucleotides expanded on each side (default:10, min:5, max:50)}
}
\value{
1/ A detailed table with the location of the detected splice sites in each
sequence and the corresponding similarity score

2/ A summary table with the number of splice sites detected in each sequence (input `meinter` function)
}
\description{
Detects potential splice sites in the proximal region of the input genomic coordinates.
The function implements the prediction model proposed by Shapiro and Senapathy (Shapiro MB, Senapathy P.
Nucleic Acids Research. 1987;15(17):7155-7174.)
}
