% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transf.R
\name{findTFBS}
\alias{findTFBS}
\title{Find putative transcription factor binding sites}
\usage{
findTFBS(bed.data, persim = 0.8, offset = 12, target = "PROMOTER",
  up.tss = 1000, down.tss = 100, mcores = NULL, tf.ID = NULL)
}
\arguments{
\item{bed.data}{A data frame containing input bed-formatted data}

\item{persim}{Minimum similarity with transcription factors consensus matrices (default:0.8, range in [0,1])}

\item{offset}{Number of nucleotides expanded on each side (default:12, min:5, max:100)}

\item{target}{Search for transcription factor binding sites on specific regions.
`PROMOTER`: selects sites located in promoter regions, `CGI`: selects sites in CpG islands,
`ALL`: No filtering is applied (time-consuming for large datasets) (default: "PROMOTER")}

\item{up.tss}{Number of nucleotides upstream transcription
start site (Only when target="PROMOTER" is set, default: 1000)}

\item{down.tss}{Number of nucleotides downstream transcription
start site (Only when target="PROMOTER" is set, default: 100)}

\item{mcores}{Number of cores to be used (default: maximum available)}

\item{tf.ID}{A vector of JASPAR transcription factors identifiers to search for (default: all)}
}
\value{
1/ Data frame containing the transcription factors identified in each sequence, their
position and binding score (input to `plotTF` function)

2/ Data frame of the detected transcription factor binding sites per sequence (input to `meinter` function)
}
\description{
Detects JASPAR's transcription factor binding sites (core collection), co-localized with input data. Both sequence strands
are examined. The analysis can be restricted to promoters (use `uptss` and `down.tss` to define promoter length, relative
to transcription start site) and CpG islands of the human genome (hg19).
}
