% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{importGEO}
\alias{importGEO}
\title{Import GEO data series in the workspace}
\usage{
importGEO(gse.acc, annotation.file)
}
\arguments{
\item{gse.acc}{A string corresponding to the accession number of the GEO data series}

\item{annotation.file}{A string corresponding to the full local path to the annotation files containing
sample grouping information}
}
\value{
1/ A bed-formatted data frame with the chromosomal coordinates and of each methylation probe and
the corresponding delta-beta values between the two groups

2/ Beta values of each sample listed in the annotation file

3/ Annotation data frame

4/ Mean beta values of group 1

5/ Mean beta values of group 2
}
\description{
Imports GEO data series. The function fetches data matrices corresponding to a pre-defined GSE identifier
and builds valid, bed-formatted dataset with delta-beta values between two sample groups, described
in a user-defined annotation file.
}
