% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{importLimma}
\alias{importLimma}
\title{Imports the results of differential methylation analyses exported by \code{limma}}
\usage{
importLimma(ltop.obj, platform = "hm450k", sortBy = "adj.P.Val")
}
\arguments{
\item{ltop.obj}{Output of \code{limma} differential analysis. The output data frame produced by the \code{limma::topTable} function.}

\item{platform}{A string corresponding to the human methylation array. Possible values are "hm27k", "hm450k" and "EPIC" (default:hm450k)}

\item{sortBy}{The criterion for selecting probes of the MArrayLM object. Possible values are: "logFC", "P.Value", "adj.P.Val"}
}
\value{
df A well-formatted data frame to be used as input to the MeinteR workflow.
}
\description{
Imports the results of differential methylation analyses exported by \code{limma}
}
