% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{isEmptyDF}
\alias{isEmptyDF}
\title{Check if data frame is empty}
\usage{
isEmptyDF(df)
}
\arguments{
\item{df}{The input data frame}
}
\value{
TRUE/FALSE (TRUE is the data frame is empty)
}
\description{
Checks if a data frame has no values
}
