% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{loadFile}
\alias{loadFile}
\title{Load input data}
\usage{
loadFile(FH)
}
\arguments{
\item{FH}{Full path of the tabular methylation data}
}
\value{
df   A data frame with the tabular methylation data
}
\description{
Loads tabular files containing methylation data. The function checks the delimiter and validates the order
of the columns (chr, start,end,score,strand).
}
