% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{loadSeqGEO}
\alias{loadSeqGEO}
\title{Reformat methylation sequencing data fetched from GEO}
\usage{
loadSeqGEO(file.path, cov = 10, chroms = NULL)
}
\arguments{
\item{file.path}{Local folder of the bed.gz file}

\item{cov}{Minimum read coverage (default:10)}

\item{chroms}{A vector of chromosome vector to be included in the analysis (default:ALL)}
}
\value{
A valid bed-formatted data frame
}
\description{
Transforms sequencing data into bed-formatted files. Valid for per sample usage.
}
