% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transf.R
\name{meinter}
\alias{meinter}
\title{Calculate the genomic index of methylation sites based on the Meinter's `find*` functions' outputs}
\usage{
meinter(bed.data, funList, weights)
}
\arguments{
\item{bed.data}{A data frame containing input bed-formatted data}

\item{funList}{List of `find*` functions outputs. At least one core function is needed to calculate
the genomic index. Valid element names of the list: `spls`-`findSpliceSites`, `altss`-`findAltSplicing`,
`ctfbs`-`findConservedTFBS`, `tfbs`-`findTFBS`, `pals`-`findPals`, `quads`-`findQuads`, `shapes`-`findShapes`}

\item{weights}{A list of positive values corresponding to feature weights [0,10]. Same elements with
`funList` list}
}
\value{
A data frame with the genomic index of the input data
}
\description{
Calculates the genomic index given a set of features pre-analysed using MeinteR's `find*` functions.
First, the function builds the local genomic signature of each site and the it calculates the genomic
index using a weighting scheme.
}
