% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{nameStudy}
\alias{nameStudy}
\title{Set a study name}
\usage{
nameStudy(study.name)
}
\arguments{
\item{study.name}{A string corresponding to the name of the study}
}
\value{
The name of the study
}
\description{
Sets a name to the analysis that appears in the exported plots.
}
