% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{plotBeta}
\alias{plotBeta}
\title{Plot scores of the input data
Generates a density plot of the score values listed in the input dataset.}
\usage{
plotBeta(bed.data)
}
\arguments{
\item{bed.data}{A data frame containing input bed-formatted data}
}
\description{
Plot scores of the input data
Generates a density plot of the score values listed in the input dataset.
}
