% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{plotCpG}
\alias{plotCpG}
\title{Plot G+C-content and observed/expected ratio.}
\usage{
plotCpG(bed.data, offset = 200)
}
\arguments{
\item{bed.data}{A data frame containing input bed-formatted data}

\item{offset}{Number of nucleotides expanded in each direction (default:200, min:20 max:1000)}
}
\value{
1/ A data frame containing the G+C content (percentage of island that is C or G)
and ratio of observed (CpG number) to expected(Number of C* Number of G/sequence length)

2/ Density plot of the G+C-content

3/ Density plot of the observed/expected ratio
}
\description{
Generates density plots of the G+C content and
observed/expected CpG ratio for the input dataset and the human genome CpG islands
}
