% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transf.R
\name{plotTF}
\alias{plotTF}
\title{Create barplot of the identified transcription factor binding sites}
\usage{
plotTF(df, topTF = 10)
}
\arguments{
\item{df}{The data frame exported by the `findTFBS` function}

\item{topTF}{Integer that sets the number of the most frequent trascription
factors to be displayed (default:10)}
}
\value{
1/ A barplot with the `topTF` most frequent transcription factors

2/ A barplot with the number of transcription factors per class

3/ A scatterplot comparing the observed and expected number of transcription
factors per class
}
\description{
Generates an overlayed barplot of the results exported by the `findTFBS` function.
The bar plot visualises the most frequent transription factors with respect
to the total number of occurrences and the number of sequences that contain
these trascription factors.
}
