\name{refFreq}
\alias{refFreq}
\docType{data}
\title{Human transcription factor frequency (data)
}
\description{
Frequency of the human conserved transcription factors in the reference genome. The frequency of each 
transcription factor is calculated using the UCSC Table Browser (tables: tfbsConsFactors, tfbsConsSites).
}
\format{
  \describe{
    \item{\code{factor}}{Name of the transcription factor}
    \item{\code{freq}}{Frequency in the human genome}
  }
}
\keyword{datasets}
