% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{reorderBed}
\alias{reorderBed}
\title{Reorder tabular methylation data to bed format}
\usage{
reorderBed(input.data, chr.col, start.col, end.col, score.col,
  strand.col = NULL)
}
\arguments{
\item{input.data}{A data frame containing input bed-formatted data}

\item{chr.col}{Column number containing the chromosome name}

\item{start.col}{Column number containing the chromosome's start position}

\item{end.col}{Column number containing the chromosome's end position}

\item{score.col}{Column number containing the methylation score values either beta or delta-beta}

\item{strand.col}{Column number containing the strand in the use data file ('+' strand is assumed if strand column is missing)}
}
\value{
A valid bed-formatted file (input of the `MeinteR::find*` functions)
}
\description{
Reorders tabular methylation data to bed-formatted files.
Compatible inputs are .txt, .csv data and other textual formats that
contain the following mandatory columns: chr, start, end and score.
}
