% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transf.R
\name{scatterConsTF}
\alias{scatterConsTF}
\title{Create a scatterplot of the identified conserved transcription factors}
\usage{
scatterConsTF(df)
}
\arguments{
\item{df}{The data frame exported by the `findConservedTFBS` function}
}
\description{
Generates a scatterplot of the results exported by the `findConservedTFBS` function. The scatterplot
illustrates the number of binding sites per transription factor relative to the expected
frequency on the reference human genome. The trascription factors with high frequency (>= 3rd quantile)
in the reference genome or to the analysed data are labeled on the scatterplot.
}
