\name{TF.class}
\alias{TF.class}
\docType{data}
\title{Transcription factor classes (data)
}
\description{
Tabular file containing the number of transcription factors per class.
}
\format{
  A data frame with 34 trascription factor classes and their corresponding number of factors in each class.
  \describe{
    \item{\code{class}}{Name of the class}
    \item{\code{Number}}{Number of factors in the class}
  }
}
\keyword{datasets}
