% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{validateBed}
\alias{validateBed}
\title{Validate format of the input bed data}
\usage{
validateBed(bed.data, omit.na = TRUE)
}
\arguments{
\item{bed.data}{A data frame containing input bed-formatted data}

\item{omit.na}{Omit rows with empty cells (default:TRUE)}
}
\value{
A well-formatted data frame
}
\description{
Validates input methylation data. Checks the presence of the chr, start, end, score columns.
If column `strand` is not set then `+` strand is assumed.
Cleans rows containing empty cells and sets numeric format to the start, end and score columns.
}
